requirements.txt
click==7.1.2
dnspython==2.0.0
email-validator==1.1.2
fastapi==0.63.0
h11==0.12.0
idna==3.1
psycopg2-binary==2.8.6
py-bcrypt==0.4
pydantic==1.7.3
python-dotenv==0.15.0
SQLAlchemy==1.3.22
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.3
